Only allow mock commands on PATH when executing a command
Before that change, the tested command had access to *all* commands that
were on the PATH by default. This made it hard to detect and write mocks
for commands that test the availabilty of a particular command with
`which <cmd>` or by probing the execution through sth. like:
`try { exec('<cmd> --version') } catch { ... }`.